// Pandem2-Prototype-FE/src/app/core/models/modelling-data.model.ts
export class ModellingScenarioParameterValue implements IModellingScenarioParameterValue {
  value: number | boolean;
  age?: ModellingModelParameterValueAgeType;
  ageContact?: ModellingModelParameterValueAgeContactType;

  constructor(data = null) {
    this.value = _.get(data, 'value');
    this.age = _.get(data, 'age');
    this.ageContact = _.get(data, 'ageContact');
  }
}

export class ModellingScenarioParameter implements IModellingScenarioParameter {
  key: string;
  values: ModellingScenarioParameterValue[];

  constructor(data = null) {
    this.key = _.get(data, 'key');
    this.values = _.get(data, 'values');
  }
}

export class ModellingAlternativeScenario implements IModellingAlternativeScenario {
  id: string;
  name: string;

  constructor(data = null) {
    this.id = _.get(data, 'id');
    this.name = _.get(data, 'name');
  }
}

export class ModellingScenario implements IModellingScenarioDataEntity {
  id: string;
  userId: string;
  modelId: string;
  name: string;
  r0: number;
  date: Date;
  description?: string;
  tags: string[];
  location: string;
  parameters: ModellingScenarioParameter[];
  sections_details?: IModellingSectionDetails[];
  exploration?: ModellingExplorationChartData[];
  is_visible?: boolean;
  alternatives?: ModellingAlternativeScenario[];

  constructor(data = null) {
    this.id = _.get(data, '_id');
    this.userId = _.get(data, 'userId');
    this.modelId = _.get(data, 'modelId');
    this.name = _.get(data, 'name');
    this.r0 = _.get(data, 'r0');
    this.date = _.get(data, 'date');
    this.description = _.get(data, 'description');
    this.tags = _.get(data, 'tags');
    this.location = _.get(data, 'location');
    this.parameters = _.get(data, 'parameters');
    this.sections_details = _.get(data, 'sections_details');
    this.exploration = _.get(data, 'exploration');
    this.is_visible = _.get(data, 'is_visible');
    this.alternatives = _.get(data, 'alternatives');
  }
}
// ...
export class ModellingExplorationChartData implements IModellingExplorationChartData {
  chart_type: 'spline' | 'column' | 'area';
  chart_plot_type: LinearLog;
  view_by: 'scenario' | 'indicator';
  view_style: ModellingViewType;
  values: string[];
  plotlines: string[];

  constructor(data = null) {
    this.chart_type = _.get(data, 'chart_type');
    this.chart_plot_type = _.get(data, 'chart_plot_type');
    this.view_by = _.get(data, 'view_by');
    this.view_style = _.get(data, 'view_style');
    this.values = _.get(data, 'values');
    this.plotlines = _.get(data, 'plotlines');
  }
}